complete crud operations

+++ app/app.js
@@ -6,6 +6,7 @@ const app = express();
 const port = process.env.PORT || 3000;
 const nodemailer = require('nodemailer');
 const path = require("path");
+const mongoose = require("mongoose");
 app.listen(port, () => console.log(`Server is running on port ${port}`));
 
 app.use(express.static(__dirname));
@@ -31,13 +32,22 @@ async function send() {
     });
 }
 
-app.use((req, res, next) => {
-    res.header('Access-Control-Allow-Origin', '*');
-    next();
-});
+// app.use((req, res, next) => {
+//     res.header('Access-Control-Allow-Origin', '*');
+//     next();
+// });
 
 app.get('/emails/:id', async (req, res) => {
     try {
+        if (isNotValidId(req)) {
+            sendBadIdResponse(res);
+            return;
+        }
+        const check = await isReqIdPresent(req);
+        if (!check) {
+            sendNotPresentIdResponse(res);
+            return;
+        }
         const documents = await emailModel.findById(req.params.id);
         res.json(documents);
     } catch (err) {
@@ -48,7 +58,36 @@ app.get('/emails/:id', async (req, res) => {
 
 app.put('/emails/:id', async (req, res) => {
     try {
+        if (isNotValidId(req)) {
+            sendBadIdResponse(res);
+            return;
+        }
+        const check = await isReqIdPresent(req);
+        if (!check) {
+            sendNotPresentIdResponse(res);
+            return;
+        }
+        const uniqueEmail = await isReqEmailUnique(req);
+        if (!uniqueEmail) {
+            sendNotUniqueEmailResponse(res);
+            return;
+        }
         const documents = await emailModel.findByIdAndUpdate(req.params.id, req.body);
+        res.json(documents);
+    } catch (err) {
+        console.error(err);
+        res.status(500).send('Server Error');
+    }
+});
+
+app.post('/emails/create', async (req, res) => {
+    try {
+        const uniqueEmail = await isReqEmailUnique(req);
+        if (!uniqueEmail) {
+            sendNotUniqueEmailResponse(res);
+            return;
+        }
+        const documents = await emailModel.create(req.body);
         res.json(documents);
     } catch (err) {
         console.error(err);
@@ -58,6 +97,15 @@ app.put('/emails/:id', async (req, res) => {
 
 app.delete('/emails/:id', async (req, res) => {
     try {
+        if (isNotValidId(req)) {
+            sendBadIdResponse(res);
+            return;
+        }
+        const check = await isReqIdPresent(req);
+        if (!check) {
+            sendNotPresentIdResponse(res);
+            return;
+        }
         await emailModel.findByIdAndDelete(req.params.id);
         res.status(204).send();
     } catch (err) {
@@ -66,7 +114,20 @@ app.delete('/emails/:id', async (req, res) => {
     }
 });
 
+app.get('/crud/new' , async (req, res) => {
+    res.render("email_create.pug");
+});
+
 app.get('/crud/:id', async (req, res) => {
+    if (isNotValidId(req)) {
+        sendBadIdResponse(res);
+        return;
+    }
+    const check = await isReqIdPresent(req);
+    if (!check) {
+        sendNotPresentIdResponse(res);
+        return;
+    }
     res.render("email_crud.pug", {doc_id: req.params.id});
 });
 
@@ -95,6 +156,32 @@ app.get('/emails.json', async (req, res) => {
 app.get('/', function (req, res) {
     res.sendFile(path.join(__dirname, '/index.html'));
 });
+
+function isNotValidId(req) {
+    return !mongoose.Types.ObjectId.isValid(req.params.id);
+}
+
+async function isReqIdPresent(req) {
+    const check = await emailModel.findById(req.params.id);
+    return check != null;
+}
+
+async function isReqEmailUnique(req) {
+    const check = await emailModel.findOne({email: req.body.email});
+    return check == null;
+}
+
+function sendNotPresentIdResponse(res) {
+    res.status(404).send("Id not found!");
+}
+
+function sendNotUniqueEmailResponse(res) {
+    res.status(400).send("Email not unique!");
+}
+
+function sendBadIdResponse(res) {
+    res.status(400).send("Bad id!");
+}
 
 
 emailModel.find().then((res) => console.log(res));
